fix: update redirects to new urls
