+++ gatsby-node.js
@@ -17,6 +17,7 @@ const CAREER_DETAILS_TEMPLATE_PATH = path.resolve(
 );
 
 const { siteMetadata } = require('./gatsby-config');
+const { createRedirects } = require('./gatsby/create-pages/create-redirects');
 const PERSONIO_JOBS_URL = 'https://satellytes.jobs.personio.de/xml';
 const PERSONIO_SHORT_DESCRIPTION_NAME = 'Kurzbeschreibung';
 const LANGUAGES = ['en', 'de'];
@@ -258,61 +259,6 @@ const createBlogPages = async ({ actions, reporter, graphql }) => {
       path: appendTrailingSlash(node.frontmatter.path),
       component: BLOG_POST_TEMPLATE_PATH,
       context: {},
-    });
-  });
-};
-
-/**
- * Sometimes we mess up URLs. Fixing them is necessary, but if Google or other
- * Link on external pages still link to the old URL they will only get the 404
- * page. That is why we need to redirect the old links to the new ones.
- */
-const createRedirects = ({ actions }) => {
-  const { createRedirect } = actions;
-
-  const redirects = [
-    {
-      fromPath: '/career/325433-senior-backend-engineer-m-w-x/',
-      toPath: '/career/325433-senior-backend-engineer-(mwx)/',
-    },
-    {
-      fromPath: '/career/325420-senior-frontend-engineer-m-w-x/',
-      toPath: '/career/325420-senior-frontend-engineer-(mwx)/',
-    },
-    {
-      fromPath: '/career/338504-frontend-engineer-w-m-x/',
-      toPath: '/career/338504-frontend-engineer-(wmx)/',
-    },
-    {
-      fromPath: '/career/325493-senior-uxui-designer-m-w-x/',
-      toPath: '/career/325493-senior-uxui-designer-(mwx)/',
-    },
-    {
-      fromPath: '/page/imprint/',
-      toPath: '/imprint/',
-    },
-  ];
-
-  redirects.forEach((redirect) => {
-    createRedirect({
-      fromPath: redirect.fromPath,
-      toPath: redirect.toPath,
-      isPermanent: true,
-      redirectInBrowser: true,
-    });
-
-    // the redirect only works for exact matches, that's why we also need to
-    // create a redirect without the trailing slash
-    const additionalFromPath = redirect.fromPath.substring(
-      0,
-      redirect.fromPath.length - 1,
-    );
-
-    createRedirect({
-      fromPath: additionalFromPath,
-      toPath: redirect.toPath,
-      isPermanent: true,
-      redirectInBrowser: true,
     });
   });
 };
